chore: make KAFKA more configurable.
